cmd/wick: core: support for exiting after receiving an event (#108)
* cmd/wick: core: support for exiting after receiving an event

Add a --exit flag to subscribe command, such that wick will exit after receiving
an event on the desired topic.

Signed-off-by: Maciej Borzecki <maciek@thing.com>

* cmd/wick: refactor auto exit after prescribed event number

Refactor how the exit after prescribed event count works. There is a problem
with synchronization across all the sessions. Since unsubscribe happens in a
deferred code, it is possible that after the event count has been reached, but
before sessions are unsubscribed, new events will be published and the code will
deadlock. Making event channel buffer larger decreases the changes of this event
occurring in practice.

Signed-off-by: Maciej Borzecki <maciek@thing.com>

Signed-off-by: Maciej Borzecki <maciek@thing.com>